+++ src/renderer/src/components/message/CopyMessageButton.tsx
@@ -8,9 +8,10 @@ import { showInfoNotification } from '@/utils/notifications';
 type Props = {
   content: string;
   className?: string;
+  alwaysShow?: boolean;
 };
 
-export const CopyMessageButton = ({ content, className }: Props) => {
+export const CopyMessageButton = ({ content, className, alwaysShow = false }: Props) => {
   const { t } = useTranslation();
   const copyToClipboard = () => {
     navigator.clipboard.writeText(content);
@@ -22,7 +23,7 @@ export const CopyMessageButton = ({ content, className }: Props) => {
       icon={<BiCopy className={`h-4 w-4 ${className}`} />}
       onClick={copyToClipboard}
       tooltip={t('messages.copy')}
-      className="opacity-0 group-hover:opacity-100"
+      className={alwaysShow ? '' : 'opacity-0 group-hover:opacity-100'}
     />
   );
 };